requirements.txt
google-genai==1.3.0
requests==2.32.3